Prevent the deletion of domains with sub-domains
This stops the accidental deletion of sub-domains within Infoblox when
the parent domain is deleted.  In order for a parent domain to be
deleted now, the sub-domains will need to be removed first.

This can be protected against with the `depends_on` directive within the
users terraform code (sub-domains referencing parent domains) but this
is easily forgotten.